Move proto-dag reader to separate file and change name
License: MIT
